requirements.txt
Django==3.2.4
python-decouple==3.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.1